Revert "switch default MockBlock to be u32"
This reverts commit b304cfd9cad3b5db2fa39e2f48b64ea6d7e77bf5.

+++ substrate/frame/system/src/tests.rs
@@ -512,7 +512,7 @@ fn deposit_event_uses_actual_weight_and_pays_fee() {
 #[test]
 fn deposit_event_topics() {
 	new_test_ext().execute_with(|| {
-		const BLOCK_NUMBER: u32 = 1;
+		const BLOCK_NUMBER: u64 = 1;
 
 		System::reset_events();
 		System::initialize(&BLOCK_NUMBER, &[0u8; 32].into(), &Default::default());
